requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
pydantic==2.12.5
PyJWT==2.11.0
requests==2.32.5
python-dotenv==1.2.1
stripe==11.5.0
motor==3.7.1
dnspython==2.8.0
email-validator==2.3.0